+++ assets/themes/default/i18n/openveo-portal-locale_en.js
@@ -55,17 +55,7 @@ angular.module('ov.locale', [], ['$provide', function($provide) {
     },
     LIVE: {
       PAGE_TITLE: 'Portal - Live',
-      TITLE: 'Live',
-      WOWZA: {
-        AUTO: 'Auto',
-        BUFFERING: 'Buffering...',
-        COUNT_DOWN: 'Time until show',
-        STREAM_ERROR: 'Stream unavailable',
-        CORS_ERROR: 'Stream unavailable (check cross-origin policy)',
-        LIVE_BUTTON: 'LIVE',
-        LIVE_BUTTON_OVER: 'Seek to LIVE',
-        LIVE_ENDED: 'This live event has ended'
-      }
+      TITLE: 'Live'
     },
     LOGIN: {
       PAGE_TITLE: 'Portal - Login',