+++ tests/client/unitTests/LiveController.js
@@ -69,50 +69,26 @@ describe('LiveController', function() {
     assert.equal($sce.getTrustedResourceUrl(scope.url), `https://console.vodalys.studio/${expectedVideoId}#?embedded=true`);
   });
 
-  it('should be able to build Wowza player', function() {
+  it('should be able to build Video.js player', function() {
     openVeoPortalSettings.live.playerType = 'wowza';
-    openVeoPortalSettings.live.wowza.playerLicenseKey = 'license-key';
     openVeoPortalSettings.live.url = 'https://wowza-example.local:1935/application-example/stream-example/playlist.m3u8';
 
-    WowzaPlayer.create = chai.spy((id, configuration) => {
-      const translate = $filter('translate');
-
-      assert.equal(id, 'wowza-player', 'Wrong player id');
-      assert.equal(configuration.license, openVeoPortalSettings.live.wowza.playerLicenseKey, 'Wrong player license');
-      assert.equal(configuration.sourceURL, `${openVeoPortalSettings.live.url}?DVR`, 'Wrong URL');
-      assert.equal(configuration.stringAuto, translate('LIVE.WOWZA.AUTO'), 'Wrong auto string');
-      assert.equal(configuration.stringBuffering, translate('LIVE.WOWZA.BUFFERING'), 'Wrong buffering string');
-      assert.equal(configuration.stringLiveLabel, translate('LIVE.WOWZA.LIVE_BUTTON'), 'Wrong live label string');
-      assert.equal(
-        configuration.stringCountdownTimerLabel,
-        translate('LIVE.WOWZA.COUNT_DOWN'),
-        'Wrong count down timer string'
-      );
-      assert.equal(
-        configuration.stringErrorStreamUnavailable,
-        translate('LIVE.WOWZA.STREAM_ERROR'),
-        'Wrong error stream unavailable string'
-      );
-      assert.equal(
-        configuration.stringErrorCORSStreamUnavailable,
-        translate('LIVE.WOWZA.CORS_ERROR'),
-        'Wrong error CORS stream unavailable string'
-      );
-      assert.equal(
-        configuration.stringLiveEventEnded,
-        translate('LIVE.WOWZA.LIVE_ENDED'),
-        'Wrong live event ended string'
-      );
-      assert.equal(
-        configuration.stringLiveSeekAlt,
-        translate('LIVE.WOWZA.LIVE_BUTTON_OVER'),
-        'Wrong live seek alt string'
-      );
+    videojs = chai.spy((id, configuration) => {
+      assert.equal(id, 'videojs-player', 'Wrong player id');
+      assert.isNotOk(configuration.autoplay, 'Unexpected player autoplay');
+      assert.ok(configuration.controls, 'Expected player controls');
+      assert.lengthOf(configuration.sources, 1, 'Expected only one player source');
+      assert.equal(configuration.sources.src, `${openVeoPortalSettings.live.url}?DVR`, 'Wrong player source URL');
+      assert.equal(configuration.sources.type, 'application/x-mpegURL', 'Wrong player source type');
+      assert.ok(configuration.liveui, 'Expected player live interface');
+      assert.isNotOk(configuration.html5.nativeControlsForTouch, 'Unexpected player native touch controls');
+      assert.isNotOk(configuration.html5.nativeAudioTracks, 'Unexpected player native audio tracks');
+      assert.isNotOk(configuration.html5.nativeVideoTracks, 'Unexpected player native video tracks');
     });
 
     initController();
 
-    WowzaPlayer.create.should.have.been.called.exactly(1);
+    videojs.should.have.been.called.exactly(1);
   });
 
 });